requirements.txt
pandas
jupyter==1.*
plotly
numpy
matplotlib
pyarrow
streamlit>=1.30
altair<6
jupyter-book==1.*
pandera
ruff
pytest
pytest-cov
pandas-gbq
google-cloud-bigquery
db-dtypes
google-cloud-bigquery-storage
